Update LICENSE
Pretty sure this project was not properly licensed before.  On Nukepedia it was listed as it is now listed here, it will be licensed in this way going forward.

To anyone who previously relied on this software to be licensed as GPL V3, this is probably inaccurate and you should update your projects accordingly.  Beyond that, I have no legal advice to give.